requirements.txt
async-timeout==5.0.1
aiohttp>=3.13.0
beautifulsoup4==4.13.4
bidict==0.23.1
blinker==1.9.0
CacheControl==0.14.3
cachelib==0.13.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cmdstanpy==1.2.5
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.2
curl_cffi==0.10.0
cycler==0.12.1
dnspython==2.7.0
et_xmlfile==2.0.0
gunicorn==21.2.0
filelock==3.18.0
finnhub-python==2.4.20
firebase-admin>=6.9.0
Flask==3.1.0
Flask-Caching==2.3.1
Flask-SocketIO==5.3.6
fonttools==4.57.0
frozendict==2.4.6
fsspec==2025.5.1
fuzzywuzzy==0.18.0
google-api-core>=2.15.0,<3.0.0
google-api-python-client>=2.169.0
google-auth>=2.23.0
google-auth-httplib2>=0.2.0
google-cloud-core>=2.4.3
google-cloud-firestore>=2.20.2
google-cloud-storage>=3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
holidays==0.71
httplib2==0.22.0
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
kaleido==0.2.1
kiwisolver==1.4.8
Levenshtein==0.27.1
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
msgpack==1.1.0
multitasking==0.0.11
narwhals==1.37.1
networkx==3.4.2
numpy==2.2.5
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pandas-datareader==0.10.0
pillow==11.2.1
platformdirs==4.3.7
plotly==6.0.1
prophet==1.1.6
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.2
python-Levenshtein==0.27.1
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
redis==6.1.0
regex==2024.11.6
reportlab==4.4.4
requests==2.32.3
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.15.3
simple-websocket==1.1.0
six==1.17.0
soupsieve==2.7
stanio==0.5.1
textblob==0.17.1
sympy==1.14.0
ta==0.11.0
threadpoolctl==3.6.0
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
Werkzeug==3.1.3
wsproto==1.2.0
yfinance==0.2.58
google-generativeai>=0.3.0
feedparser==6.0.11
pytrends==4.9.1
selenium==4.15.2
webdriver-manager==4.0.1
vaderSentiment==3.3.2
praw==7.7.1